requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Chemistry
rdkit>=2022.03.1

# Configuration
pyyaml>=6.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Progress bar
tqdm>=4.62.0

# Optional: for faster sparse matrix operations
# scikit-learn>=1.0.0
